+++ res/pieces/pieces-service/src/main/java/com/pieces/service/vo/CropInfo.java
@@ -9,16 +9,16 @@ public class CropInfo {
 
     private String imgUrl;        // your image path (the one we recieved after successfull upload)
 
-    private Integer imgInitW;    // your image original width (the one we recieved after upload)
-    private Integer imgInitH;    // your image original height (the one we recieved after upload)
+    private Double imgInitW;    // your image original width (the one we recieved after upload)
+    private Double imgInitH;    // your image original height (the one we recieved after upload)
 
-    private Integer imgW;        // your new scaled image width
-    private Integer imgH;    // your new scaled image height
-    private Integer imgX1;        // top left corner of the cropped image in relation to scaled image
-    private Integer imgY1;    // top left corner of the cropped image in relation to scaled image
+    private Double imgW;        // your new scaled image width
+    private Double imgH;    // your new scaled image height
+    private Double imgX1;        // top left corner of the cropped image in relation to scaled image
+    private Double imgY1;    // top left corner of the cropped image in relation to scaled image
 
-    private Integer cropW;        // cropped image width
-    private Integer cropH;    // cropped image height
+    private Double cropW;        // cropped image width
+    private Double cropH;    // cropped image height
 
     public String getImgUrl() {
         return imgUrl;
@@ -28,67 +28,67 @@ public class CropInfo {
         this.imgUrl = imgUrl;
     }
 
-    public Integer getImgInitW() {
+    public Double getImgInitW() {
         return imgInitW;
     }
 
-    public void setImgInitW(Integer imgInitW) {
+    public void setImgInitW(Double imgInitW) {
         this.imgInitW = imgInitW;
     }
 
-    public Integer getImgInitH() {
+    public Double getImgInitH() {
         return imgInitH;
     }
 
-    public void setImgInitH(Integer imgInitH) {
+    public void setImgInitH(Double imgInitH) {
         this.imgInitH = imgInitH;
     }
 
-    public Integer getImgW() {
+    public Double getImgW() {
         return imgW;
     }
 
-    public void setImgW(Integer imgW) {
+    public void setImgW(Double imgW) {
         this.imgW = imgW;
     }
 
-    public Integer getImgH() {
+    public Double getImgH() {
         return imgH;
     }
 
-    public void setImgH(Integer imgH) {
+    public void setImgH(Double imgH) {
         this.imgH = imgH;
     }
 
-    public Integer getImgX1() {
+    public Double getImgX1() {
         return imgX1;
     }
 
-    public void setImgX1(Integer imgX1) {
+    public void setImgX1(Double imgX1) {
         this.imgX1 = imgX1;
     }
 
-    public Integer getImgY1() {
+    public Double getImgY1() {
         return imgY1;
     }
 
-    public void setImgY1(Integer imgY1) {
+    public void setImgY1(Double imgY1) {
         this.imgY1 = imgY1;
     }
 
-    public Integer getCropW() {
+    public Double getCropW() {
         return cropW;
     }
 
-    public void setCropW(Integer cropW) {
+    public void setCropW(Double cropW) {
         this.cropW = cropW;
     }
 
-    public Integer getCropH() {
+    public Double getCropH() {
         return cropH;
     }
 
-    public void setCropH(Integer cropH) {
+    public void setCropH(Double cropH) {
         this.cropH = cropH;
     }
 }
